Added trip model to the project.
This trip model is called from the trip view controller. Now the
project has the MVC structure.
I still have to add the feedback from the model to the controller. This
should be done by using notifications. For now the long and lat
displayed in the view are not showing the data due to this issue.